Moves query code from list_cmd to didicated module.
